+++ src/http/httpd.c
@@ -12,16 +12,16 @@ void httpd_client_ping (httpd_client_t * http, int timeout);
 int httpd_client_reread (httpd_client_t * http);
 
 void httpd_client_clean (httpd_client_t * http) {
-  if (master.debug & DEBUG_MEMORY) printf ("httpd clean %d\n", http->c.sockfd);
+  if (http->debug & DEBUG_MEMORY) printf ("httpd clean %d\n", http->c.sockfd);
   if (http->file_path) free ((void*)http->file_path);
   if (http->post_sep) free ((void*)http->post_sep);
 
   if (http->file_fd) {
-    if (master.debug & DEBUG_SYSCALL) printf ("  close file_fd %d\n", http->file_fd);
+    if (http->debug & DEBUG_SYSCALL) printf ("  close file_fd %d\n", http->file_fd);
     close (http->file_fd);
   }
   if (!HIN_HTTPD_WORKER_PREFORKED && http->post_fd) {
-    if (master.debug & DEBUG_SYSCALL) printf ("  close post_fd %d\n", http->post_fd);
+    if (http->debug & DEBUG_SYSCALL) printf ("  close post_fd %d\n", http->post_fd);
     close (http->post_fd); // TODO cgi worker needs to keep this
   }
   http->file_fd = http->post_fd = 0;
@@ -40,8 +40,9 @@ void httpd_client_clean (httpd_client_t * http) {
 }
 
 int httpd_client_start_request (httpd_client_t * http) {
-  if (master.debug & DEBUG_PROTO) printf ("httpd request begin %d\n", http->c.sockfd);
   http->state = HIN_REQ_HEADERS | (http->state & HIN_REQ_ENDING);
+  http->debug = master.debug;
+  if (http->debug & DEBUG_PROTO) printf ("httpd request begin %d\n", http->c.sockfd);
 
   hin_client_t * server = (hin_client_t*)http->c.parent;
   hin_server_data_t * data = (hin_server_data_t*)server->parent;
@@ -61,7 +62,7 @@ int httpd_client_finish_request (httpd_client_t * http) {
   if (http->state & (HIN_REQ_POST)) return 0;
 
   int keep = (http->peer_flags & HIN_HTTP_KEEPALIVE) && ((http->state & HIN_REQ_ENDING) == 0);
-  if (master.debug & DEBUG_PROTO) printf ("httpd request done %d %s\n", http->c.sockfd, keep ? "keep" : "close");
+  if (http->debug & DEBUG_PROTO) printf ("httpd request done %d %s\n", http->c.sockfd, keep ? "keep" : "close");
 
   int hin_server_finish_callback (httpd_client_t * client);
   hin_server_finish_callback (http);
@@ -85,7 +86,7 @@ static int httpd_client_close_callback (hin_buffer_t * buffer, int ret) {
     printf ("httpd client close callback error: %s\n", strerror (-ret));
     return -1;
   }
-  if (master.debug & DEBUG_PROTO) printf ("httpd close %d\n", http->c.sockfd);
+  if (http->debug & DEBUG_PROTO) printf ("httpd close %d\n", http->c.sockfd);
   if (http->read_buffer && http->read_buffer != buffer) {
     hin_buffer_clean (http->read_buffer);
     http->read_buffer = NULL;
@@ -98,7 +99,7 @@ static int httpd_client_close_callback (hin_buffer_t * buffer, int ret) {
 int httpd_client_shutdown (httpd_client_t * http) {
   if (http->state & HIN_REQ_ENDING) return -1;
   http->state |= HIN_REQ_ENDING;
-  if (master.debug & DEBUG_SOCKET) printf ("httpd shutdown %d\n", http->c.sockfd);
+  if (http->debug & DEBUG_SOCKET) printf ("httpd shutdown %d\n", http->c.sockfd);
 
 #if 0
   struct linger sl;
@@ -121,7 +122,7 @@ int httpd_client_shutdown (httpd_client_t * http) {
 
 static int httpd_client_buffer_close_callback (hin_buffer_t * buffer) {
   httpd_client_t * http = (httpd_client_t*)buffer->parent;
-  if (master.debug & DEBUG_PROTO) printf ("httpd shutdown buffer %d\n", http->c.sockfd);
+  if (http->debug & DEBUG_PROTO) printf ("httpd shutdown buffer %d\n", http->c.sockfd);
   httpd_client_shutdown (http);
   return 0;
 }