+++ src/http/httpd_file.c
@@ -147,7 +147,7 @@ int httpd_send_file (httpd_client_t * http, hin_cache_item_t * item, hin_buffer_
 
   header (buf, "\r\n");
 
-  if (master.debug & DEBUG_RW) printf ("responding '\n%.*s'\n", buf->count, buf->ptr);
+  if (http->debug & DEBUG_RW) printf ("responding %d '\n%.*s'\n", http->c.sockfd, buf->count, buf->ptr);
 
   hin_pipe_write (pipe, buf);
   hin_pipe_start (pipe);